Homework2/Component1/CloudFunctions/GetFeedbackAnalysis/requirements.txt
functions-framework==3.*
google-cloud-firestore>=2.15.0
google-cloud-storage==2.16.0